Add files via upload
This "Tests" project was made to test my current abilities in python as of 07.12.2021. What I wanted to do was to have the program receive information from the user and administer a test specific to them. Immediately we jump into it. The information asked, which is needed would be the name and age of the user. With that information inputted the user will be asked for conformation. If there were any errors made then it would loop back to get the correct information.  Once again it will ask for confirmation on the new data that was inputted. If in the off chance the information is still wrong then the user will be looped back to go through the process until the mistake is corrected. Once all is well then the respective test will be administered. The test is multiple choice and consist of three simple questions. When all answers are in you will be presented with your score.