Harvard Course Example Improved
In one segment of week two of the online course a small program is made to get the average of a set of scores. Harvard only presented an example for three elements. Taking it into my own hands, I made it possible to get the average for any amount of elements the user requires.